Started doing something
I'm using the Buffer class from the practicals as a simple queue data structure for the orders. I added a getSize() method to it in order to avoid making a foreach every time I needed the number of elements in the queue.

I started changing the conventions regarding the naming of method paramenters.

Created a MyCook class from the SampleCook class supplied in the code, I'll expand it as I go if needed.

Need to figure out what to use for the pizza delivery chain (another queue?) and for the ingredients shelves.

All of this would be easier if I had a tasksheet outlining what should be done in the tasks.